+++ cmd/aws-sso/cache_cmd.go
@@ -20,9 +20,15 @@ package main
 
 import (
 	"fmt"
+
+	"github.com/synfinatic/aws-sso-cli/internal/awsconfig"
+	"github.com/synfinatic/aws-sso-cli/internal/url"
 )
 
-type CacheCmd struct{}
+type CacheCmd struct {
+	NoConfigCheck bool `kong:"help='Disable automatic ~/.aws/config updates'"`
+	Threads       int  `kong:"help='Override number of threads for talking to AWS'"`
+}
 
 func (cc *CacheCmd) Run(ctx *RunContext) error {
 	s, err := ctx.Settings.GetSelectedSSO(ctx.Cli.SSO)
@@ -44,6 +50,17 @@ func (cc *CacheCmd) Run(ctx *RunContext) error {
 	err = ctx.Settings.Cache.Save(true)
 	if err != nil {
 		return fmt.Errorf("Unable to save role cache: %s", err.Error())
+	}
+
+	// should we update our config??
+	if !ctx.Cli.Cache.NoConfigCheck && ctx.Settings.AutoConfigCheck {
+		if ctx.Settings.ConfigProfilesUrlAction != url.ConfigProfilesUndef {
+			action, _ := url.NewAction(string(ctx.Settings.ConfigProfilesUrlAction))
+			err := awsconfig.UpdateAwsConfig(ctx.Settings, action, "", true, false)
+			if err != nil {
+				log.Errorf("Unable to auto-update aws config file: %s", err.Error())
+			}
+		}
 	}
 
 	return nil